Move classic Option/Result to dedicated dirs.
