+++ src/Option.ts
@@ -1,0 +1,8 @@
+export {
+    FlatmapFn,
+    MayRecoveryFn,
+    OptionBase,
+    Option,
+    createSome,
+    createNone,
+} from './ClassOption/Option';
