+++ src/Result.ts
@@ -1,0 +1,8 @@
+export {
+    FlatmapOkFn,
+    FlatmapErrFn,
+    ResultBase,
+    Result,
+    createOk,
+    createErr,
+} from './ClassResult/Result';
