+++ tools/pkg_files.json
@@ -3,6 +3,18 @@
     "CHANGELOG_OLD.md": null,
     "LICENSE.MIT": null,
     "README.md": null,
+    "cjs/ClassOption/Option.d.ts": {
+        "private": true
+    },
+    "cjs/ClassOption/Option.js": {
+        "private": true
+    },
+    "cjs/ClassResult/Result.d.ts": {
+        "private": true
+    },
+    "cjs/ClassResult/Result.js": {
+        "private": true
+    },
     "cjs/Maybe/ErrorMessage.d.ts": null,
     "cjs/Maybe/ErrorMessage.js": null,
     "cjs/Maybe/Maybe.d.ts": null,
@@ -199,6 +211,18 @@
     "docs/idiom/README.md": null,
     "docs/public_api_list.md": null,
     "docs/wrapper_objects.md": null,
+    "esm/ClassOption/Option.d.ts": {
+        "private": true
+    },
+    "esm/ClassOption/Option.mjs": {
+        "private": true
+    },
+    "esm/ClassResult/Result.d.ts": {
+        "private": true
+    },
+    "esm/ClassResult/Result.mjs": {
+        "private": true
+    },
     "esm/Maybe/ErrorMessage.d.ts": null,
     "esm/Maybe/ErrorMessage.mjs": null,
     "esm/Maybe/Maybe.d.ts": null,
@@ -391,6 +415,24 @@
     "esm/shared/Function.mjs": null,
     "esm/shared/Mutable.d.ts": null,
     "esm/shared/Mutable.mjs": null,
+    "lib/ClassOption/Option.d.ts": {
+        "private": true
+    },
+    "lib/ClassOption/Option.js": {
+        "private": true
+    },
+    "lib/ClassOption/Option.mjs": {
+        "private": true
+    },
+    "lib/ClassResult/Result.d.ts": {
+        "private": true
+    },
+    "lib/ClassResult/Result.js": {
+        "private": true
+    },
+    "lib/ClassResult/Result.mjs": {
+        "private": true
+    },
     "lib/Maybe/ErrorMessage.d.ts": null,
     "lib/Maybe/ErrorMessage.js": null,
     "lib/Maybe/ErrorMessage.mjs": null,
